Made pkg a required argument instead of optional.
Also separated the the path from the pkg so that only the src-dir
needs to be specified. The other dirs are generated using the package
name.

Had to update the calling syntax on some items slightly. Most notably,
the constructors no longer accept their sql-items anymore. You now
have to call the appropriate "add-method" in most cases.

Signed-off-by: Jonas Kalderstam <jonas@kalderstam.se>